Change nvcuvid link to dyn + hint for FFmpeg
